facenet_svm_version/requirements.txt
opencv-python
numpy
mtcnn
keras-facenet
scikit-learn
joblib
tqdm
streamlit
Pillow
